Add support for using the signal feature
This will allow users to start kraken and listen for a signal before
starting the chaos scenarios. This is helpful in cases where we want
to coordinate between two frameworks and run chaos when the state of
cluster is at a desired state or when we want to inject chaos when a
particular load test starts for example.